A mostly-rewrite of lambda composer logic around a proper DSl and a split-layer library resolver mechanism lambdas are now first-class citizens in the API, 'compose' is no longer needed, and types are stricter all around.

+++ metagen-lang/src/main/java/io/virtdata/parser/MetagenBuilder.java
@@ -122,7 +122,7 @@ public class MetagenBuilder extends MetagenBaseListener {
 
     @Override
     public void exitFloatValue(MetagenParser.FloatValueContext ctx) {
-        calls.peek().addFunctionArg(new FloatArg(Double.valueOf(ctx.getText())));
+        calls.peek().addFunctionArg(new FloatArg(Float.valueOf(ctx.getText())));
     }
 
     @Override
